# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kayak Simulation Distributed Services API
  version: 1.0.0
  description: |
    API specification for the Kayak-style distributed system. The platform exposes REST
    endpoints for user management, listings search, bookings, billing, admin operations,
    analytics, and the FastAPI-based agentic recommendation service. All synchronous
    operations occur over HTTP while asynchronous updates are published on Kafka topics
    and relayed to clients via WebSockets.
  contact:
    name: Kayak Simulation Platform Team
    email: eng@kayak-sim.local
servers:
  - url: https://api.kayak-sim.prod
    description: Production API Gateway
  - url: https://api.staging.kayak-sim.local
    description: Staging API Gateway
  - url: http://localhost:8080
    description: Local development gateway
...
paths:
...
  /api/v1/hotels/{hotelId}:
    parameters:
      - name: hotelId
        in: path
        required: true
        schema:
          type: string
          pattern: '^[A-Z0-9\\-]{3,32}$'
    get:
      tags: [Listings]
      summary: Get hotel details
      operationId: getHotel
      responses:
        '200':
          description: Hotel details payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/v1/cars/{carId}:
    parameters:
      - name: carId
        in: path
        required: true
        schema:
          type: string
          pattern: '^[A-Z0-9\\-]{3,32}$'
    get:
      tags: [Listings]
      summary: Get car details
      operationId: getCar
      responses:
        '200':
          description: Car details payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Car'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/bookings:
    get:
      tags: [Bookings]
      summary: Search bookings
      operationId: searchBookings
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: bookingType
          in: query
          schema:
            $ref: '#/components/schemas/BookingType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated booking list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
    post:
      tags: [Bookings]
      summary: Create booking
      description: Creates a booking and kicks off downstream Kafka workflows.
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateRequest'
      responses:
        '201':
          description: Booking created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Conflicting booking for the same inventory slice.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/bookings/{bookingId}:
    parameters:
      - name: bookingId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Bookings]
      summary: Get booking
      operationId: getBooking
      responses:
        '200':
          description: Booking details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [Bookings]
      summary: Update booking status
      operationId: updateBookingStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdateRequest'
      responses:
        '200':
          description: Booking status updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination (1-indexed).
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
      description: Number of items to return per page.
...
  responses:
    ValidationError:
      description: Input validation failed; see details array.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    PaginationMetadata:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 200
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        correlationId:
          type: string
          format: uuid
          description: Trace ID for support teams.
        details:
          type: array
          items:
            type: string
          description: Additional error context.
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: array
              description: Field level validation failures.
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  rejectedValue:
                    description: The value that failed validation.
    Address:
      type: object
      required: [line1, city, state, zipCode]
      properties:
        line1:
          type: string
          maxLength: 255
        line2:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 120
        state:
          $ref: '#/components/schemas/USStateCode'
        zipCode:
          type: string
          pattern: '^\\d{5}(-\\d{4})?$'
          description: Supports '#####' or '#####-####'.
...
    Review:
      type: object
      required: [id, userId, listingType, listingId, rating, title, body, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        listingType:
          type: string
          enum: [flight, hotel, car]
        listingId:
          type: string
        rating:
          type: number
          minimum: 1
          maximum: 5
        title:
          type: string
          maxLength: 120
        body:
          type: string
          maxLength: 2000
        createdAt:
          type: string
          format: date-time
...
    FlightClass:
      type: string
      enum: [economy, premium_economy, business, first]
...
    Hotel:
      type: object
      required:
        - id
        - name
        - address
        - starRating
        - totalRooms
        - pricePerNight
        - amenities
      properties:
        id:
          type: string
        name:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        neighborhood:
          type: string
        starRating:
          type: number
          minimum: 0
          maximum: 5
        totalRooms:
          type: integer
          minimum: 1
        roomTypes:
          type: array
          items:
            $ref: '#/components/schemas/RoomType'
        pricePerNight:
          $ref: '#/components/schemas/Money'
        amenities:
          type: array
          items:
            type: string
        rating:
          type: number
          minimum: 0
          maximum: 5
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        images:
          type: array
          items:
            type: string
            format: uri
...
    RoomType:
      type: object
      required: [code, description, capacity, pricePerNight, refundable]
      properties:
        code:
          type: string
        description:
          type: string
        capacity:
          type: integer
          minimum: 1
        beds:
          type: integer
          minimum: 1
        pricePerNight:
          $ref: '#/components/schemas/Money'
        refundable:
          type: boolean
    Car:
      type: object
      required:
        - id
        - provider
        - model
        - carType
        - transmission
        - seats
        - pricePerDay
        - availabilityStatus
      properties:
        id:
          type: string
        provider:
          type: string
        model:
          type: string
        year:
          type: integer
          minimum: 1990
        carType:
          $ref: '#/components/schemas/CarType'
        transmission:
          type: string
          enum: [automatic, manual]
        seats:
          type: integer
          minimum: 1
        pricePerDay:
          $ref: '#/components/schemas/Money'
        rating:
          type: number
          minimum: 0
          maximum: 5
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        availabilityStatus:
          type: string
          enum: [available, reserved, maintenance]
...
    CarType:
      type: string
      enum: [suv, sedan, compact, luxury, convertible, van, truck]
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: float
          minimum: 0
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: USD
    BookingType:
      type: string
      enum: [flight, hotel, car]
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, COMPLETED, CANCELLED, FAILED]
    Booking:
      type: object
      required:
        - id
        - userId
        - bookingType
        - status
        - createdAt
        - updatedAt
        - itinerary
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        bookingType:
          $ref: '#/components/schemas/BookingType'
        status:
          $ref: '#/components/schemas/BookingStatus'
        itinerary:
          $ref: '#/components/schemas/Itinerary'
        price:
          $ref: '#/components/schemas/Money'
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookingCreateRequest:
      type: object
      required:
        - userId
        - bookingType
        - itinerary
        - quoteExpiresAt
      properties:
        userId:
          type: string
          format: uuid
        bookingType:
          $ref: '#/components/schemas/BookingType'
        itinerary:
          $ref: '#/components/schemas/Itinerary'
        quoteId:
          type: string
        quoteExpiresAt:
          type: string
          format: date-time
        paymentMethodId:
          type: string
          format: uuid
        metadata:
          type: object
          additionalProperties: true
    BookingUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/BookingStatus'
        cancellationReason:
          type: string
          maxLength: 500
    BookingListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Booking'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
    Itinerary:
      type: object
      required: [segments]
      properties:
        segments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ItinerarySegment'
        guests:
          type: array
          items:
            $ref: '#/components/schemas/Guest'
    ItinerarySegment:
      type: object
      required: [type, referenceId, startDateTime, endDateTime]
      properties:
        type:
          type: string
          enum: [flight, hotel, car]
        referenceId:
          type: string
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        fareClass:
          $ref: '#/components/schemas/FlightClass'
        roomTypeCode:
          type: string
        carType:
          $ref: '#/components/schemas/CarType'
        seatAssignments:
          type: array
          items:
            type: string
        amenityCodes:
          type: array
          items:
            type: string
    Guest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          pattern: '^\\+1-\\d{3}-\\d{3}-\\d{4}$'
...
    USStateCode:
      type: string
      enum:
        - AL
        - AK
        - AZ
        - AR
        - CA
        - CO
        - CT
        - DE
        - FL
        - GA
        - HI
        - ID
        - IL
        - IN
        - IA
        - KS
        - KY
        - LA
        - ME
        - MD
        - MA
        - MI
        - MN
        - MS
        - MO
        - MT
        - NE
        - NV
        - NH
        - NJ
        - NM
        - NY
        - NC
        - ND
        - OH
        - OK
        - OR
        - PA
        - RI
        - SC
        - SD
        - TN
        - TX
        - UT
        - VT
        - VA
        - WA
        - WV
        - WI
        - WY
      description: Two-letter postal abbreviation for US states.
...